move unix code to unix only location
